Migrate fjs/types/sorted_set to .f.mjs, split types into types.ts
Splits the SortedSet<T> type into a sibling types.ts, and converts the
runtime implementation (union, intersect, has) to JSDoc-typed
module.f.mjs/proof.f.mjs under the stage-1 TypeScript-to-mjs migration.
Updates the three dependents (fjs/types/range_map, fjs/types/byte_set,
fjs/fsm) to import runtime values from module.f.mjs and the SortedSet
type from types.ts.
