+++ fjs/fsm/module.f.ts
@@ -7,7 +7,8 @@ import type { List } from '../types/list/types.ts'
 import { equal, isEmpty, fold, toArray, scan, foldScan, empty as emptyList } from '../types/list/module.f.mjs'
 import type { StringMap } from '../types/object/types.ts'
 import { toRangeMap, union as byteSetUnion, one, empty, range, type ByteSet } from '../types/byte_set/module.f.ts'
-import { intersect, type SortedSet, union as sortedSetUnion } from '../types/sorted_set/module.f.ts'
+import { intersect, union as sortedSetUnion } from '../types/sorted_set/module.f.mjs'
+import type { SortedSet } from '../types/sorted_set/types.ts'
 import type { RangeMap, Properties, RangeMapArray, Entry } from '../types/range_map/types.ts'
 import { merge, get as rangeMapGet } from '../types/range_map/module.f.mjs'
 import type { Fold, Scan } from '../types/function/operator/types.ts'
