+++ fjs/types/byte_set/module.f.ts
@@ -5,7 +5,7 @@
  */
 import { compose } from '../function/module.f.mjs'
 import type { RangeMap } from '../range_map/types.ts'
-import type { SortedSet } from '../sorted_set/module.f.ts'
+import type { SortedSet } from '../sorted_set/types.ts'
 import { reverse, countdown, flat, map } from '../list/module.f.mjs'
 
 export type ByteSet = bigint
