+++ fjs/types/range_map/proof.f.ts
@@ -2,7 +2,8 @@ import type { RangeMapArray, Properties, RangeMap } from './types.ts'
 import { get, merge, fromRange, rangeMap } from './module.f.mjs'
 import { stringify, type Unknown } from '../../media/json/module.f.ts'
 import { sort } from '../object/module.f.mjs'
-import { union, type SortedSet } from '../sorted_set/module.f.ts'
+import { union } from '../sorted_set/module.f.mjs'
+import type { SortedSet } from '../sorted_set/types.ts'
 import { equal, toArray } from '../list/module.f.mjs'
 import { strictEqual } from '../function/operator/module.f.mjs'
 import { cmp } from '../string/module.f.mjs'
