+++ fjs/types/sorted_set/types.ts
@@ -1,0 +1,7 @@
+/**
+ * Type-level API for the sorted-set module.
+ *
+ * @module
+ */
+
+export type SortedSet<T> = readonly T[]
